# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Glorp Garage API
  version: "1"
  description: >
    Remote control plane for Glorp Garage: create workspaces and run coding
    agents over an API-key-secured HTTP/WS API. All endpoints are served under
    the stable `/api/v1` prefix (and, for backwards compatibility, at the root).
    Auth is required whenever Garage is bound to a non-loopback host.


    Multi-tenancy: an admin key can provision isolated **namespaces** (one per
    user/tenant) and mint namespace-bound keys. A tenant key transparently scopes
    every call to its own namespace; an admin key can act in any namespace by
    sending `X-Glorp-Namespace: <id>`. Requests with no namespace resolve to the
    built-in `default` namespace (the legacy, single-tenant layout).
servers:
  - url: "{endpoint}/api/v1"
    variables:
      endpoint:
        default: http://127.0.0.1:4271
        description: Base URL of the Garage server.
security:
  - bearerAuth: []
...
paths:
...
  /sessions:
    get:
      tags: [sessions]
      summary: List sessions (scoped to the resolved namespace)
      parameters: [{ $ref: "#/components/parameters/NamespaceHeader" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/SessionDto" }
                  total: { type: integer }
    post:
      tags: [sessions]
      summary: Create a session (in the resolved namespace)
      parameters: [{ $ref: "#/components/parameters/NamespaceHeader" }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSessionInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionDto" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer glsk_…`. For the WebSocket event stream
        (`/api/v1/sessions/{id}/events`), which cannot set headers from a
        browser, pass the key as an `?api_key=glsk_…` query parameter instead.
  parameters:
...
    NamespaceHeader:
      name: X-Glorp-Namespace
      in: header
      required: false
      schema: { type: string }
      description: >
        Admin keys only: act inside this namespace. Tenant keys are bound to
        their own namespace and a mismatching value is rejected (403). Omit to
        use the key's own namespace (or `default`).
...
  schemas:
...
    CreateSessionInput:
      type: object
      properties:
        workspaceId: { type: string }
        workspace: { type: string, description: Absolute host path. }
        template: { type: string }
        params: { type: object, additionalProperties: { type: string } }
        provider: { type: string }
        model: { type: string }
        profileId: { type: string }
        permissionMode: { type: string, enum: [normal, auto, bypass] }
    SessionDto:
      type: object
      properties:
        id: { type: string }
        state: { type: string, enum: [provisioning, idle, busy, error, destroyed] }
        workspace: { type: string }
        workspace_id: { type: [string, "null"] }
        title: { type: [string, "null"] }
        model_label: { type: [string, "null"] }
        permission_mode: { type: string, enum: [normal, auto, bypass] }
        busy: { type: boolean }
        turn_count: { type: integer }
        tokens_in: { type: integer }
        tokens_out: { type: integer }
        error: { type: [string, "null"] }
        ws_url: { type: string, description: WebSocket URL for the event stream. }
...